Refactor D3D12 generator files and helper types
