+++ src/ComputeSharp.SourceGenerators/ComputeShaderDescriptorGenerator.ResourceDescriptorRanges.Syntax.cs
@@ -8,32 +8,8 @@ namespace ComputeSharp.SourceGenerators;
 partial class ComputeShaderDescriptorGenerator
 {
     /// <inheritdoc/>
-    partial class LoadDispatchMetadata
+    partial class ResourceDescriptorRanges
     {
-        /// <summary>
-        /// Writes the <c>ConstantBufferSize</c> property.
-        /// </summary>
-        /// <param name="info">The input <see cref="ShaderInfo"/> instance with gathered shader info.</param>
-        /// <param name="writer">The <see cref="IndentedTextWriter"/> instance to write into.</param>
-        public static void WriteConstantBufferSizeSyntax(ShaderInfo info, IndentedTextWriter writer)
-        {
-            writer.WriteLine("/// <inheritdoc/>");
-            writer.WriteGeneratedAttributes(GeneratorName);
-            writer.WriteLine($"readonly int global::ComputeSharp.Descriptors.IComputeShaderDescriptor<{info.Hierarchy.Hierarchy[0].QualifiedName}>.ConstantBufferSize => {info.ConstantBufferSizeInBytes};");
-        }
-
-        /// <summary>
-        /// Writes the <c>IsStaticSamplerRequired</c> property.
-        /// </summary>
-        /// <param name="info">The input <see cref="ShaderInfo"/> instance with gathered shader info.</param>
-        /// <param name="writer">The <see cref="IndentedTextWriter"/> instance to write into.</param>
-        public static void WriteIsStaticSamplerRequiredSyntax(ShaderInfo info, IndentedTextWriter writer)
-        {
-            writer.WriteLine("/// <inheritdoc/>");
-            writer.WriteGeneratedAttributes(GeneratorName);
-            writer.WriteLine($"readonly bool global::ComputeSharp.Descriptors.IComputeShaderDescriptor<{info.Hierarchy.Hierarchy[0].QualifiedName}>.IsStaticSamplerRequired => {info.IsSamplerUsed.ToString().ToLowerInvariant()};");
-        }
-
         /// <summary>
         /// Writes the <c>ResourceDescriptorRanges</c> property.
         /// </summary>
